# src/vaultspec_a2a/api/internal.py
"""Internal endpoints for worker <-> gateway communication.

The worker process communicates with the gateway via two channels:

1. **WebSocket** (``/internal/ws``) -- streaming path where the worker
   pushes ``WorkerEventEnvelope`` and ``HeartbeatMessage`` JSON frames.
2. **HTTP POST** (``/internal/events``, ``/internal/heartbeat``) -- preferred
   path that avoids the need for a WebSocket client library in the worker.
   The ``WorkerBridge`` in ``vaultspec_a2a.worker.ipc`` uses this approach.

The gateway exposes ``/internal/health`` for readiness probes.
"""

from __future__ import annotations

import json
import logging
import time
from typing import Any

from fastapi import (
    APIRouter,
    Depends,
    Header,
    HTTPException,
    Request,
    WebSocket,
    WebSocketDisconnect,
)

from ..control.config import settings
from ..control.event_handlers import (
    _handle_execution_state_event,
    relay_event,
)
from ..thread.snapshots import normalize_wire_event_type
from ..utils import BearerVerdict, verify_internal_bearer

__all__ = ["internal_router"]

logger = logging.getLogger(__name__)


def _validate_event_envelope(
    thread_id: str,
    payload: dict[str, Any],
    *,
    context: str,
) -> None:
    """Reject malformed worker event payloads with a client-visible error."""
    if not thread_id or not payload:
        raise HTTPException(
            status_code=422,
            detail=f"Malformed {context}: thread_id and payload are required",
        )


async def _verify_internal_token(
    authorization: str | None = Header(None),
) -> None:
    """Verify bearer token for internal IPC endpoints.

    Skipped when settings.internal_token is None **and** the environment
    is DEVELOPMENT.  In production/staging/testing, a missing token is a
    configuration error. Delegates the rule to the shared IPC bearer verifier.
    """
    verdict, detail = verify_internal_bearer(
        authorization,
        token=settings.internal_token,
        environment=settings.environment,
    )
    if verdict is BearerVerdict.MISCONFIGURED:
        raise HTTPException(status_code=500, detail=detail)
    if verdict is BearerVerdict.UNAUTHORIZED:
        raise HTTPException(status_code=401, detail=detail)


internal_router = APIRouter(
    prefix="/internal",
    tags=["internal"],
    dependencies=[Depends(_verify_internal_token)],
)


async def _relay_single_event(
    thread_id: str,
    payload: dict[str, Any],
    *,
    agg: Any,
    session_factory: Any,
    drain_gate: Any = None,
    transport: str = "http",
) -> None:
    """Aggregate and relay a single worker event.

    Shared by all three ingest paths (WS, HTTP POST, HTTP batch) to
    avoid copy-pasting the aggregator sync → relay_event sequence.

    *drain_gate* is the process-wide run-admission gate seated on ``app.state``;
    it travels to the terminal handler, which releases the run from it, exactly
    as *agg* and *session_factory* travel to their handlers.
    """
    payload = normalize_wire_event_type(payload)
    if payload.get("type") == "execution_state_projection":
        await _handle_execution_state_event(
            thread_id, payload, session_factory=session_factory
        )
        return

    if agg is not None:
        agg.relay_payload(thread_id, payload)
        agg.sync_worker_event(thread_id, payload)
    else:
        logger.warning(
            "No relay target available -- dropping event for %s",
            thread_id,
            extra={
                "thread_id": thread_id,
                "event_type": str(payload.get("event_type", payload.get("type", ""))),
                "transport": transport,
                "action": "relay_drop_event",
            },
        )
    await relay_event(
        thread_id,
        payload,
        aggregator=agg,
        session_factory=session_factory,
        drain_gate=drain_gate,
    )


async def _relay_worker_event(websocket: WebSocket, msg: dict, raw: str) -> None:
    """Relay a single worker event to WS clients and update aggregator/DB."""
    thread_id = msg.get("thread_id", "")
    payload = msg.get("payload", {})
    if not thread_id or not payload:
        logger.warning(
            "Malformed worker event envelope: %s",
            raw[:200],
            extra={
                "thread_id": thread_id,
                "event_type": "",
                "message_type": str(msg.get("type", "")),
                "transport": "ws",
                "frame_size": len(raw),
            },
        )
        return
    session_factory = getattr(websocket.app.state, "db_session_factory", None)
    agg = getattr(websocket.app.state, "aggregator", None)
    # Read the seated gate rather than get-or-creating it: a gate that has never
    # been seated has admitted nothing, so there is nothing to release.
    drain_gate = getattr(websocket.app.state, "drain_gate", None)
    await _relay_single_event(
        thread_id,
        payload,
        agg=agg,
        session_factory=session_factory,
        drain_gate=drain_gate,
        transport="ws",
    )


@internal_router.websocket("/ws")
async def worker_ws_endpoint(websocket: WebSocket) -> None:
    """Internal WebSocket -- receives events and heartbeats from the worker."""
    # WebSocket routes bypass router-level Depends(), so enforce the same worker
    # IPC bearer rule the HTTP endpoints use through the one shared authority. The
    # gateway-facing internal channel accepts only the worker IPC credential, never
    # the attach credential; a non-OK verdict refuses the connection.
    verdict, _ = verify_internal_bearer(
        websocket.headers.get("authorization"),
        token=settings.internal_token,
        environment=settings.environment,
    )
    if verdict is not BearerVerdict.OK:
        await websocket.close(code=1008, reason="Unauthorized")
        return

    await websocket.accept()
    logger.info("Worker connected to internal WS")

    # Store reference so supervisor can check connectivity
    websocket.app.state.worker_ws = websocket
    websocket.app.state.worker_last_heartbeat_ts = time.monotonic()

    try:
        while True:
            raw = await websocket.receive_text()
            # Reject oversized frames (1 MB) to prevent memory exhaustion.
            if len(raw) > settings.internal_max_frame_bytes:
                logger.warning(
                    "Dropping oversized internal WS frame (%d bytes)", len(raw)
                )
                continue
            msg = json.loads(raw)
            msg_type = msg.get("type", "")

            match msg_type:
                case "heartbeat":
                    websocket.app.state.worker_last_heartbeat_ts = time.monotonic()
                    websocket.app.state.worker_active_threads = msg.get(
                        "active_threads", []
                    )
                    logger.debug(
                        "Worker heartbeat: %d active threads",
                        len(msg.get("active_threads", [])),
                        extra={
                            "message_type": msg_type,
                            "active_thread_count": len(msg.get("active_threads", [])),
                            "transport": "ws",
                        },
                    )

                case "event":
                    await _relay_worker_event(websocket, msg, raw)

                case _:
                    logger.warning(
                        "Unknown internal WS message type: %s",
                        msg_type,
                        extra={
                            "message_type": msg_type,
                            "transport": "ws",
                            "frame_size": len(raw),
                        },
                    )

    except WebSocketDisconnect:
        logger.warning("Worker disconnected from internal WS")
        websocket.app.state.worker_ws = None


@internal_router.get("/health")
async def internal_health() -> dict[str, str]:
    """Readiness probe -- confirms the internal API is accepting connections."""
    return {"status": "ok", "service": "gateway"}


# ---------------------------------------------------------------------------
# HTTP POST endpoints -- preferred IPC path (avoids WS client dependency)
# ---------------------------------------------------------------------------


@internal_router.post("/events")
async def receive_worker_event(request: Request) -> dict[str, str]:
    """Receive a single event from the worker and relay to browser clients.

    The ``WorkerBridge`` in the worker process POSTs here instead of
    sending a WebSocket frame.  The payload format matches
    ``WorkerEventEnvelope``.
    """
    # Reject oversized payloads (1 MB) on internal HTTP path.
    content_length = request.headers.get("content-length")
    try:
        if (
            content_length is not None
            and int(content_length) > settings.internal_max_http_body_bytes
        ):
            raise HTTPException(status_code=413, detail="Payload too large (max 1 MB)")
    except ValueError as e:
        raise HTTPException(
            status_code=400,
            detail="Invalid Content-Length header",
        ) from e

    body: dict[str, Any] = await request.json()
    thread_id: str = body.get("thread_id", "")
    payload: dict[str, Any] = body.get("payload", {})
    _validate_event_envelope(thread_id, payload, context="worker event POST")

    agg = getattr(request.app.state, "aggregator", None)
    if agg is None:
        raise HTTPException(
            status_code=503,
            detail="No relay target available -- gateway not ready",
        )

    await _relay_single_event(
        thread_id,
        payload,
        agg=agg,
        session_factory=getattr(request.app.state, "db_session_factory", None),
        drain_gate=getattr(request.app.state, "drain_gate", None),
    )
    return {"status": "ok"}


@internal_router.post("/events/batch")
async def receive_worker_event_batch(request: Request) -> dict[str, str]:
    """Receive a batch of events from the worker.

    The ``WorkerBridge`` accumulates events for a short interval then
    POSTs them as a single ``{"events": [...]}`` payload.  Each entry
    has the same shape as a single-event envelope (``thread_id`` +
    ``payload``).
    """
    content_length = request.headers.get("content-length")
    # Allow larger batches: 4 MB limit for batch payloads.
    try:
        if (
            content_length is not None
            and int(content_length) > settings.internal_max_http_body_bytes * 4
        ):
            raise HTTPException(
                status_code=413,
                detail="Payload too large (max 4 MB)",
            )
    except ValueError as e:
        raise HTTPException(
            status_code=400,
            detail="Invalid Content-Length header",
        ) from e

    body: dict[str, Any] = await request.json()
    events: list[dict[str, Any]] = body.get("events", [])

    # Sort events by worker-side monotonic timestamp to preserve causal order
    # even if the batch was assembled out of order.
    events.sort(key=lambda e: e.get("ts", 0.0))

    agg = getattr(request.app.state, "aggregator", None)
    if agg is None:
        raise HTTPException(
            status_code=503,
            detail="No relay target available -- gateway not ready",
        )

    session_factory = getattr(request.app.state, "db_session_factory", None)
    drain_gate = getattr(request.app.state, "drain_gate", None)

    for idx, evt in enumerate(events):
        thread_id = evt.get("thread_id", "")
        payload = evt.get("payload", {})
        try:
            _validate_event_envelope(
                thread_id, payload, context=f"worker event batch entry {idx}"
            )
        except HTTPException as exc:
            raise HTTPException(status_code=422, detail=exc.detail) from None

    for evt in events:
        thread_id = evt.get("thread_id", "")
        payload = evt.get("payload", {})
        await _relay_single_event(
            thread_id,
            payload,
            agg=agg,
            session_factory=session_factory,
            drain_gate=drain_gate,
        )

    return {"status": "ok"}


@internal_router.post("/heartbeat")
async def receive_worker_heartbeat(request: Request) -> dict[str, str]:
    """Receive a heartbeat from the worker.

    Updates ``app.state`` so the gateway can monitor worker
    liveness without a persistent WebSocket connection.
    """
    body: dict[str, Any] = await request.json()
    request.app.state.worker_last_heartbeat_ts = time.monotonic()
    request.app.state.worker_active_threads = body.get("active_threads", [])
    logger.debug(
        "Worker heartbeat (HTTP): %d active threads",
        len(body.get("active_threads", [])),
        extra={
            "message_type": str(body.get("type", "")),
            "active_thread_count": len(body.get("active_threads", [])),
            "transport": "http",
        },
    )
    return {"status": "ok"}